admin seperarted

+++ controllers/invoiceItemsController.js
@@ -5,9 +5,9 @@ exports.addInvoiceItem = async (req, res) => {
   try {
     
     const [result] = await pool.execute(
-      `INSERT INTO invoice_items (item_name, description, item_code, sale_price) 
-       VALUES (?,?,?,?)`,
-      [item_name, description, item_code, sale_price]
+      `INSERT INTO invoice_items (item_name, description, item_code, sale_price,created_by) 
+       VALUES (?,?,?,?,?)`,
+      [item_name, description, item_code, sale_price, req.user.id]
     );
     res.send({ message: 'Invoice Items created', status:200, id:result.insertId });
   } catch (err) {
@@ -18,8 +18,8 @@ exports.addInvoiceItem = async (req, res) => {
 // READ ALL
 exports.getAllInvoiceItems = async (req, res) => {
   try {
-    const [rows] = await pool.execute('SELECT * FROM invoice_items');
-    res.json({message:"Customers fetched successfully", data:rows ,status:200});
+    const [rows] = await pool.execute('SELECT * FROM invoice_items WHERE created_by = ?', [req.user.id]);
+    res.json({message:"Invoice items fetched successfully", data:rows ,status:200});
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -29,9 +29,9 @@ exports.getAllInvoiceItems = async (req, res) => {
 exports.getInvoiceItemById = async (req, res) => {
   const { id } = req.params;
   try {
-    const [rows] = await pool.execute('SELECT * FROM invoice_items WHERE id = ?', [id]);
-    if (rows.length === 0) return res.status(404).json({ message: 'Customer not found' });
-    res.json({ message: 'Customer found', data: rows[0], status:200 });
+    const [rows] = await pool.execute('SELECT * FROM invoice_items WHERE id = ? AND created_by = ?', [id, req.user.id]);
+    if (rows.length === 0) return res.status(404).json({ message: 'Item not found' });
+    res.json({ message: 'Item found', data: rows[0], status:200 });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -45,10 +45,10 @@ exports.updateInvoiceItem = async (req, res) => {
     const [result] = await pool.execute(
       `UPDATE invoice_items SET 
         item_name = ?, description =?, item_code =?, sale_price =?
-       WHERE id = ?`,
-      [item_name, item_category, description, item_code, sale_price, id]
+       WHERE id = ? AND created_by = ?`,
+      [item_name, item_category, description, item_code, sale_price, id, req.user.id]
     );
-    res.json({ message: 'Customer updated',status:200 });
+    res.json({ message: 'Invoice item updated',status:200 });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -58,8 +58,8 @@ exports.updateInvoiceItem = async (req, res) => {
 exports.deleteInvoiceItem = async (req, res) => {
   const { id } = req.params;
   try {
-    const [result] = await pool.execute('DELETE FROM invoice_items WHERE id = ?', [id]);
-    res.json({ message: 'Customer deleted',status:200 });
+    const [result] = await pool.execute('DELETE FROM invoice_items WHERE id = ? AND created_by = ?', [id, req.user.id]);
+    res.json({ message: 'Item deleted',status:200 });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
